+++ semcore/data-table/CHANGELOG.md
@@ -12,9 +12,11 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 - RawData to `CellRenderProps`.
 - `line-height` for `Head.Column` items.
 - `overflow=hidden` for `CellWrapper`.
+- `textAlign` property to all column cells.
 
 ### Fixed
 
+- `z-index` of expanded accordions and spinner container.
 - Usage of `changeSortSize` property with multiple columns.
 - Accordion content expands with visual issues when clicking on a merged cell.
 - Incorrect type for `selectedRows/onSelectedRowsChange` when `uniqueRowKey` type is not a string.